Week1 week2 programs added

+++ Week1JavaPrograms/src/com/utility/Utility.java
@@ -36,7 +36,7 @@ public class Utility {
 	   }
 	   public static float getFloat()
 	   {
-		   int floatValue=scanner.nextInt();
+		   float floatValue=scanner.nextFloat();
 		   return floatValue;
 	   }
 	   public static double getDouble()
